add README.md, partial rework of guilddata (additionally there will no longer be a "guild not in database error", the guild will be added to the database on the fly instead if that is the case), add time to various console logs, move discord command to information category and fix embed, add confirmation to kick, ban, and unban commands, rename package.json 'name' field to 'Thebotcat', fix typo for client.on message and voiceStateUpdate error
lock command partially broken, and mute and settings commands fully broken, among several other commands including music ones, additionally now that commands can be turned off with guilddata they have to actually be disabled as that's not programmed in yet

+++ common/handlers/extras/message.js
@@ -1,5 +1,6 @@
 module.exports = [
   msg => {
+    if (props.feat.version == 'canary') return;
     if (msg.channel.id == '738599826765250632') {
       msg.delete();
       client.channels.cache.get('738593863958003756').send(`${msg.author.tag}: ${msg.content}`);
